requirements.txt
dotenv
pycryptodome